containers/load_to_bq/scripts/requirements.txt
pandas
mysql-connector-python
google-cloud-storage
google-cloud-bigquery
